refactor(task): split operations.go into domain-focused operation files
Moved all I/O types and Register* functions from the monolithic operations.go
into four focused files: task_operations.go (CRUD/field/bulk/activation),
template_operations.go (templates/recurring), query_operations.go (queries),
and nlp_operations.go (natural language). No logic changes.
